\name{all_combinations}
\alias{all_combinations}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculating all possible permutations
}
\description{
This function calculates all possible permutations of order 2 in a set of order k.
}
\usage{
all_combinations(k)
}

\arguments{
  \item{k}{
Order of set. A positive integer.
}
}

\value{
A matrix.
}
\keyword{internal}
