\name{check_more_push}
\alias{check_more_push}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Helper function for Smith Normal Form calculation.
}
\description{
Checks if the diagonal entries are ordered properly.
}
\usage{
check_more_push(D)
}

\arguments{
  \item{D}{
Diagonal of a matrix that shall be turned into Smith Normal Form.
}
}
\value{
TRUE/FALSE
}




\seealso{
\code{\link{smith}}, \code{\link{push_down}}
}

\keyword{internal}
